+++ src/main/java/net/jasper/mod/gui/PlayerAutomaHUD.java
@@ -1,6 +1,7 @@
 package net.jasper.mod.gui;
 
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
+import net.jasper.mod.automation.MenuPrevention;
 import net.jasper.mod.automation.PlayerRecorder;
 import net.jasper.mod.gui.option.PlayerAutomaOptionsScreen;
 import net.jasper.mod.mixins.InGameHudDimensions;
@@ -13,7 +14,7 @@ import net.minecraft.text.Text;
 /**
  * Little HUD for Playerautoma to display current state of player recorder
  */
-public class HUDState {
+public class PlayerAutomaHUD {
 
     public enum ShowHUDOption {
         NOTHING,
@@ -31,11 +32,11 @@ public class HUDState {
             };
         }
 
-        public static HUDState.ShowHUDOption fromString(String s) {
-            return HUDState.ShowHUDOption.valueOf(s.toUpperCase());
+        public static PlayerAutomaHUD.ShowHUDOption fromString(String s) {
+            return PlayerAutomaHUD.ShowHUDOption.valueOf(s.toUpperCase());
         }
 
-        public static Text toText(HUDState.ShowHUDOption opt) {
+        public static Text toText(PlayerAutomaHUD.ShowHUDOption opt) {
             return switch(opt) {
                 case NOTHING -> Text.translatable("playerautoma.option.hudShow.nothing");
                 case TEXT -> Text.translatable("playerautoma.option.hudShow.text");
@@ -64,11 +65,11 @@ public class HUDState {
             };
         }
 
-        public static HUDState.Position fromString(String s) {
-            return HUDState.Position.valueOf(s.toUpperCase());
+        public static PlayerAutomaHUD.Position fromString(String s) {
+            return PlayerAutomaHUD.Position.valueOf(s.toUpperCase());
         }
 
-        public static Text toText(HUDState.Position opt) {
+        public static Text toText(PlayerAutomaHUD.Position opt) {
             return switch (opt) {
                 case TOP_LEFT -> Text.translatable("playerautoma.option.hudPosition.topLeft");
                 case TOP_RIGHT -> Text.translatable("playerautoma.option.hudPosition.topRight");
@@ -87,7 +88,7 @@ public class HUDState {
             // Get the longest text offset of the state if text is displayed
             TextRenderer textRenderer = MinecraftClient.getInstance().textRenderer;
             int textOffset = 0;
-            HUDState.ShowHUDOption opt = PlayerAutomaOptionsScreen.showHudOption.getValue();
+            PlayerAutomaHUD.ShowHUDOption opt = PlayerAutomaOptionsScreen.showHudOption.getValue();
             if (opt == ShowHUDOption.TEXT || opt == ShowHUDOption.TEXT_AND_ICON) {
                 for (PlayerRecorder.State s : PlayerRecorder.State.values()) {
                     textOffset = Math.max(textOffset, textRenderer.getWidth(s.getText()));
@@ -108,8 +109,12 @@ public class HUDState {
     public static void register() {
         HudRenderCallback.EVENT.register((context, tickDelta) -> {
             MinecraftClient client = MinecraftClient.getInstance();
+
+            // Renders only if active
+            MenuPrevention.renderIcon(context);
+
             ShowHUDOption showOffHud = PlayerAutomaOptionsScreen.showHudOption.getValue();
-            if (showOffHud == ShowHUDOption.NOTHING  || client.currentScreen != null) {
+            if (showOffHud == ShowHUDOption.NOTHING) {
                 return;
             }
             // Get/Calc guiScale