+++ src/main/java/net/jasper/mod/util/HUDTextures.java
@@ -1,0 +1,17 @@
+package net.jasper.mod.util;
+
+import net.jasper.mod.PlayerAutomaClient;
+import net.minecraft.util.Identifier;
+
+/**
+ * Namespace to hold all identifiers used by Playerautoma
+ */
+public class HUDTextures {
+    public static final Identifier BLOCK_MENU_ICON = new Identifier(PlayerAutomaClient.MOD_ID, "textures/gui/recorder_icons/block_menu.png");
+    public static final Identifier REPLAYING_ICON = new Identifier(PlayerAutomaClient.MOD_ID, "textures/gui/recorder_icons/replaying.png");
+    public static final Identifier RECORDING_ICON = new Identifier(PlayerAutomaClient.MOD_ID, "textures/gui/recorder_icons/recording.png");
+    public static final Identifier PAUSING_ICON = new Identifier(PlayerAutomaClient.MOD_ID, "textures/gui/recorder_icons/pausing.png");
+    public static final Identifier IDLE_ICON = new Identifier(PlayerAutomaClient.MOD_ID, "textures/gui/recorder_icons/idle.png");
+    public static final Identifier REFRESH_ICON = new Identifier(PlayerAutomaClient.MOD_ID, "textures/gui/recorder_icons/refresh.png");
+
+}