+++ src/main/java/net/jasper/mod/util/data/Recording.java
@@ -3,13 +3,16 @@ package net.jasper.mod.util.data;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Main-Class for storing all data required for a recording
  */
 public class Recording implements Serializable {
     public record RecordEntry(
-            List<Boolean> keysPressed,
+            List<String> keysPressed,
+            Map<String, Integer> timesPressed,
+            List<String> modifiers,
             LookingDirection lookingDirection,
             int slotSelection,
             SlotClick slotClicked,