+++ src/main/java/net/jasper/mod/util/keybinds/Constants.java
@@ -1,6 +1,7 @@
 package net.jasper.mod.util.keybinds;
 
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
+import net.jasper.mod.automation.MenuPrevention;
 import net.jasper.mod.automation.PlayerRecorder;
 import net.jasper.mod.gui.ModMenu;
 import net.jasper.mod.gui.RecordingSelector;
@@ -14,6 +15,10 @@ import org.lwjgl.glfw.GLFW;
  */
 public class Constants {
 
+    public static final String CTRL = "ctrl";
+    public static final String SHIFT = "shift";
+    public static final String ALT = "alt";
+
     private static final String KEYBINDING_CATEGORY = "PlayerAutoma";
 
     private static final String[] translations = {
@@ -25,7 +30,8 @@ public class Constants {
             "playerautoma.keys.storeRecording",
             "playerautoma.keys.loadRecording",
             "playerautoma.keys.pauseReplay",
-            "playerautoma.keys.openMenu"
+            "playerautoma.keys.openMenu",
+            "toBackground"
     };
 
     protected static final int AMOUNT_KEYBINDS = translations.length;
@@ -39,10 +45,12 @@ public class Constants {
             new KeyBinding(translations[5], InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_U, KEYBINDING_CATEGORY),
             new KeyBinding(translations[6], InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_I, KEYBINDING_CATEGORY),
             new KeyBinding(translations[7], InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_B, KEYBINDING_CATEGORY),
-            new KeyBinding(translations[8], InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_O, KEYBINDING_CATEGORY)
+            new KeyBinding(translations[8], InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_O, KEYBINDING_CATEGORY),
+            new KeyBinding(translations[9], InputUtil.Type.KEYSYM, GLFW.GLFW_KEY_0, KEYBINDING_CATEGORY)
     };
 
     public static final KeyBinding STOP_REPLAY = bindings[3];
+    public static final KeyBinding PREVENT_MENU = bindings[9];
 
     private static final Runnable[] callbackMethods = {
             PlayerRecorder::startRecord,
@@ -53,7 +61,8 @@ public class Constants {
             RecordingStorer::open,
             RecordingSelector::open,
             PlayerRecorder::togglePauseReplay,
-            ModMenu::open
+            ModMenu::open,
+            MenuPrevention::toggleBackgroundPrevention
     };
 
     protected static KeyBind[] defaultKeybinds = new KeyBind[AMOUNT_KEYBINDS];